requirements.txt
Django==4.1
djangorestframework==3.14.0
matplotlib==3.6.2
matplotlib-inline==0.1.6
numpy==1.24.0
pandas==1.5.2
Pillow==9.3.0
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
squarify==0.4.3